+++ pom.xml
@@ -11,7 +11,7 @@
     <groupId>com.example</groupId>
     <artifactId>HakatonSpring</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>HakatonSpring</name>
+    <name>HackathonSpring</name>
     <description>Demo project for Spring Boot</description>
 
     <properties>
