+++ src/test/java/com/hackathon/ApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.HakatonSpring;
+package com.hackathon;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class HakatonSpringApplicationTests {
+class ApplicationTests {
 
 	@Test
 	void contextLoads() {
